requirements.txt
Flask==2.3.2
pandas==2.0.3
tqdm==4.64.1
requests==2.31.0
azure-storage-blob==12.15.0
pymysql==1.1.0
requests
